+++ R/namespace.R
@@ -1,2 +1,3 @@
 #' @import chk
+#' @importFrom stats dbinom dlnorm dnorm dpois dnbinom
 NULL